requirements.txt
Django==5.1
djangorestframework==3.15.2
django-rest-knox==5.0.2
environs==11.0.0